fix: Resolve remaining CI failures in ruff, GCC, and clang-tidy
- Remove unused imports (contextmanager, Elevation, MagicMock) flagged by ruff F401
- Widen fd_link buffer to 512 in port.c to fix GCC format-truncation
- Fix narrowing conversions in gpud collector.c and main.c
- Parenthesize macro arguments in gpud nvml.c for bugprone-macro-parentheses

+++ ebpf/gpud/collector.c
@@ -34,7 +34,7 @@ static void get_process_name(unsigned int pid, char *name, size_t len)
     snprintf(path, sizeof(path), "/proc/%u/comm", pid);
     f = fopen(path, "r");
     if (f) {
-        if (fgets(name, len, f)) {
+        if (fgets(name, (int)len, f)) {
             /* Remove trailing newline */
             char *nl = strchr(name, '\n');
             if (nl)
